+++ lib/system.nim
@@ -909,6 +909,34 @@ else:
   else:
     proc reset*[T](obj: var T) {.magic: "Reset", noSideEffect.}
 
+when defined(nimHasDefault):
+  proc isDefault*[T: not seq](a: T): bool {.inline.} =
+    ## returns whether `a` is equal to its default value
+    runnableExamples:
+      doAssert "".isDefault
+      doAssert not "a".isDefault
+      doAssert (-0.0).isDefault
+      doAssert not @[0].isDefault
+      doAssert [0].isDefault # whereas [0].len != 0
+
+      type Kind = enum kUnknown, kRed, kGreen
+      doAssert kUnknown.isDefault
+
+      type Foo1 = ref object
+      type Foo2 = object
+      doAssert not Foo1().isDefault
+      doAssert Foo2().isDefault
+
+    a == default(T)
+
+template isDefault*(a: string): bool =
+  ## overloaded for efficiency
+  a.len == 0
+
+template isDefault*(a: seq): bool =
+  ## overloaded for efficiency
+  a.len == 0
+
 proc setLen*[T](s: var seq[T], newlen: Natural) {.
   magic: "SetLengthSeq", noSideEffect.}
   ## Sets the length of seq `s` to `newlen`. ``T`` may be any sequence type.